FISH-12821: restructuring modules to run MicroProfile Telemetry TCKs
