+++ MicroProfile-OpenTelemetry/MicroProfile-OpenTelemetry-JVM-Metrics-Application/pom.xml
@@ -40,70 +40,29 @@
   ~  holder.
   ~
   -->
-
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
+
     <parent>
-        <groupId>fish.payara.microprofile</groupId>
-        <artifactId>tck-suite-parent</artifactId>
+        <groupId>fish.payara.microprofile.telemetry</groupId>
+        <artifactId>MicroProfile-OpenTelemetry</artifactId>
         <version>1.1-SNAPSHOT</version>
     </parent>
-    
-    <packaging>pom</packaging>
-    
-    <groupId>fish.payara.microprofile.telemetry</groupId>
-    <artifactId>MicroProfile-OpenTelemetry</artifactId>
 
-    <properties>
-        <microprofile.telemetry.version>2.1</microprofile.telemetry.version>
-        <mptck.suite>${basedir}/src/test/resources/tck-suite.xml</mptck.suite>
-        <payara.executable>./asadmin</payara.executable>
-    </properties>
+    <artifactId>MicroProfile-OpenTelemetry-JVM-Metrics-Application</artifactId>
+    <name>MicroProfile-OpenTelemetry-JVM-Metrics-Application</name>
+    
+    <packaging>jar</packaging>
 
     <dependencies>
         <dependency>
-            <groupId>jakarta.platform</groupId>
-            <artifactId>jakarta.jakartaee-web-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.microprofile.config</groupId>
-            <artifactId>microprofile-config-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.microprofile.rest.client</groupId>
-            <artifactId>microprofile-rest-client-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.glassfish.jersey.core</groupId>
-            <artifactId>jersey-client</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.glassfish.jersey.inject</groupId>
-            <artifactId>jersey-hk2</artifactId>
+            <groupId>org.eclipse.microprofile.telemetry</groupId>
+            <artifactId>microprofile-telemetry-metrics-tck</artifactId>
+            <version>${microprofile.telemetry.version}</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
-
-    <modules>
-        <module>MicroProfile-OpenTelemetry-Tracing-Logging</module>
-        <module>MicroProfile-OpenTelemetry-JVM-Metrics-Runtime</module>
-        <module>MicroProfile-OpenTelemetry-JVM-Metrics-Application</module>
-    </modules>
-
-    <profiles>
-        <profile>
-            <id>windows-profile</id>
-            <activation>
-                <os>
-                    <family>Windows</family>
-                </os>
-            </activation>
-            <properties>
-                <payara.executable>${payara.home}${file.separator}bin${file.separator}asadmin.bat</payara.executable>
-            </properties>
-        </profile>
-    </profiles>
+    
 </project>
