+++ MicroProfile-OpenTelemetry/MicroProfile-OpenTelemetry-Tracing-Logging/src/test/java/fish/payara/microprofile/telemetry/tracing/tck/ArquillianExtension.java
@@ -2,7 +2,7 @@
  *
  *  DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  *
- *  Copyright (c) 2023-2026 Payara Foundation and/or its affiliates. All rights reserved.
+ *  Copyright (c) 2026 Payara Foundation and/or its affiliates. All rights reserved.
  *
  *  The contents of this file are subject to the terms of either the GNU
  *  General Public License Version 2 only ("GPL") or the Common Development
