+++ MicroProfile-OpenTelemetry/MicroProfile-OpenTelemetry-Tracing-Logging/src/test/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
